# Reducing rotations based on bounding box constraint