--- a/python/libraries/resizing.py
+++ b/python/libraries/resizing.py
@@ -1,5 +1,6 @@
-import numpy as np
 from typing import Generator
+
+import numpy as np
 
 
 def crop_cube(cube: np.ndarray) -> np.ndarray:
@@ -23,6 +24,53 @@ def crop_cube(cube: np.ndarray) -> np.ndarray:
     return cube
 
 
+def fix_axis(cube):
+    """
+    Rotate cube to make boundary sizes in sorted order.
+
+    Ex : if input cube.shape = (4, 1, 2) this method rotate it to be (1, 2, 4)
+
+    Parameters:
+    cube (np.array): 3D Numpy byte array where 1 values indicate polycube positions
+
+    Returns:
+    np.array: Cropped 3D Numpy byte array equivalent to cube, but with no zero padding
+
+    """
+    if cube.shape == tuple(sorted(cube.shape, reverse=True)):
+        return cube
+
+    if cube.shape == tuple(sorted(cube.shape)):
+        return np.rot90(cube, 1, (0, 2))
+
+    if cube.shape[0] < cube.shape[2] < cube.shape[1]:
+        return np.rot90(np.rot90(cube, 1, (0, 2)), 1, (1, 2))
+
+    if cube.shape[0] == cube.shape[2]:
+        if cube.shape[0] < cube.shape[1]:
+            return np.rot90(cube, 1, (0, 1))
+        else:
+            return np.rot90(cube, 1, (1, 2))
+
+    if cube.shape[0] == cube.shape[1] or cube.shape[1] == cube.shape[2]:
+        if cube.shape[0] < cube.shape[2]:
+            return cube
+        else:
+            return np.rot90(cube, 1, (0, 2))
+
+    if cube.shape[2] < cube.shape[0] < cube.shape[1]:
+        return np.rot90(cube, 1, (0, 1))
+
+    if cube.shape[1] < cube.shape[0] < cube.shape[2]:
+        return np.rot90(np.rot90(cube, 1, (0, 1)), 1, (1, 2))
+
+    if cube.shape[1] < cube.shape[2] < cube.shape[0]:
+        return np.rot90(cube, 1, (2, 1))
+
+    print("error", cube.shape)
+    exit(2)
+
+
 def expand_cube(cube: np.ndarray) -> Generator[np.ndarray, None, None]:
     """
     Expands a polycube by adding single blocks at all valid locations.
@@ -41,16 +89,16 @@ def expand_cube(cube: np.ndarray) -> Generator[np.ndarray, None, None]:
     output_cube = np.array(cube)
 
     xs, ys, zs = cube.nonzero()
-    output_cube[xs+1, ys, zs] = 1
-    output_cube[xs-1, ys, zs] = 1
-    output_cube[xs, ys+1, zs] = 1
-    output_cube[xs, ys-1, zs] = 1
-    output_cube[xs, ys, zs+1] = 1
-    output_cube[xs, ys, zs-1] = 1
+    output_cube[xs + 1, ys, zs] = 1
+    output_cube[xs - 1, ys, zs] = 1
+    output_cube[xs, ys + 1, zs] = 1
+    output_cube[xs, ys - 1, zs] = 1
+    output_cube[xs, ys, zs + 1] = 1
+    output_cube[xs, ys, zs - 1] = 1
 
     exp = (output_cube ^ cube).nonzero()
 
     for (x, y, z) in zip(exp[0], exp[1], exp[2]):
         new_cube = np.array(cube)
         new_cube[x, y, z] = 1
-        yield crop_cube(new_cube)
+        yield fix_axis(crop_cube(new_cube))