--- a/python/libraries/rotation.py
+++ b/python/libraries/rotation.py
@@ -4,7 +4,7 @@ from typing import Generator
 
 def all_rotations(polycube: np.ndarray) -> Generator[np.ndarray, None, None]:
     """
-    Calculates all rotations of a polycube.
+    Calculates rotations of a polycube when bounding box size in all axes are equal(Ex (3, 3, 3)).
 
     Adapted from https://stackoverflow.com/questions/33190042/how-to-calculate-all-24-rotations-of-3d-array.
     This function computes all 24 rotations around each of the axis x,y,z. It uses numpy operations to do this, to avoid unecessary copies.
@@ -36,3 +36,47 @@ def all_rotations(polycube: np.ndarray) -> Generator[np.ndarray, None, None]:
     # rotate about axis 2, 8 rotations about axis 1
     yield from single_axis_rotation(np.rot90(polycube, axes=(0, 1)), (0, 2))
     yield from single_axis_rotation(np.rot90(polycube, -1, axes=(0, 1)), (0, 2))
+
+
+def one_diff_rot(cube, diff_axis, equal_axes):
+    """
+    Calculates rotations of a polycube when bounding box size in two axes are equal, but one is different. (Ex (2, 2, 3)).
+    Only 8 rotations can be done without breaking the bounding box.
+
+    Parameters:
+    polycube (np.array): 3D Numpy byte array where 1 values indicate polycube positions
+    diff_axis (integer): Axis which has different size
+    equal_axes (integer array): Axes which have same size
+
+    Returns:
+    generator(np.array): Yields new rotations of this cube about all axes
+
+    """
+    for _ in range(0, 4):
+        yield cube
+        cube = np.rot90(cube, 1, equal_axes)
+
+    cube = np.rot90(cube, 2, (diff_axis, equal_axes[0]))
+
+    for _ in range(0, 4):
+        yield cube
+        cube = np.rot90(cube, 1, equal_axes)
+
+
+def all_diff_rot(cube):
+    """
+    Calculates rotations of a polycube when bounding box size in all axes are different. (Ex (1, 4, 6)).
+    Only 4 rotations can be done without breaking the bounding box.
+
+    Parameters:
+    polycube (np.array): 3D Numpy byte array where 1 values indicate polycube positions
+
+    Returns:
+    generator(np.array): Yields new rotations of this cube about all axes
+
+    """
+    yield cube
+    yield np.rot90(cube, 2, (0, 1))
+    yield np.rot90(cube, 2, (0, 2))
+    yield np.rot90(cube, 2, (1, 2))
+